requirements.txt
groq
PyMuPDF
dotenv
pytest
Pillow
fastapi
uvicorn 
python-multipart